+++ image-viewer/src/cursor.c
@@ -22,7 +22,7 @@ void cursor_center(cursor_t *cursor, image_region_t region) {
 }
 
 bool cursor_move(cursor_t *cursor, image_region_t region, direction_t direction, int16_t amount) {
-  uint16_t x = cursor->x, y = cursor->y;
+  int32_t x = cursor->x, y = cursor->y;
   direction_move_xy(direction, &x, &y, amount);
 
   if (x < region.x) {
