+++ image-viewer/src/direction.c
@@ -1,6 +1,6 @@
 #include "direction.h"
 
-void direction_move_xy(direction_t direction, uint16_t *x, uint16_t *y, int16_t amount) {
+void direction_move_xy(direction_t direction, int32_t *x, int32_t *y, int16_t amount) {
   switch (direction) {
   case LEFT:
     *x -= amount;
